requirements.txt
pandas
nltk
matplotlib
psycopg2-binary
sqlalchemy